app/requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.3
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
whitenoise==6.5.0
